[BC Break] Moves symfony and pimple routers under the Router namespace
