+++ src/Bernard/Router/ContainerAwareRouter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bernard\Symfony;
+namespace Bernard\Router;
 
 use Bernard\Envelope;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -8,7 +8,7 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 /**
  * @package Bernard
  */
-class ContainerAwareRouter extends \Bernard\Router\SimpleRouter
+class ContainerAwareRouter extends SimpleRouter
 {
     /**
      * @param ContainerInterface $container
